requirements.txt
anyio==3.6.2
attrs==22.2.0
bcrypt==4.0.1
certifi==2022.12.7
cffi==1.15.1
click==8.1.3
cryptography==39.0.1
dnspython==2.3.0
ecdsa==0.18.0
email-validator==1.3.1
exceptiongroup==1.1.0
fastapi==0.91.0
greenlet==2.0.2
h11==0.14.0
httpcore==0.16.3
httptools==0.5.0
httpx==0.23.3
idna==3.4
importlib-metadata==6.0.0
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
mysql-connector-python==8.0.32
orjson==3.8.6
packaging==23.0
passlib==1.7.4
pluggy==1.0.0
protobuf==3.20.3
pyasn1==0.4.8
pycparser==2.21
pydantic==1.10.4
pytest==7.2.1
python-dotenv==0.21.1
python-jose==3.3.0
python-multipart==0.0.5
PyYAML==6.0
rfc3986==1.5.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.3
starlette==0.24.0
tomli==2.0.1
typing_extensions==4.4.0
ujson==5.7.0
uvicorn==0.20.0
uvloop==0.17.0
watchfiles==0.18.1
websockets==10.4
zipp==3.13.0
